requirements.txt
flask==3.1.0
gunicorn==23.0.0
anthropic==0.42.0
python-dotenv==1.0.1
bcrypt==4.2.1
flask-limiter==3.8.0
